+++ src/builders/GCleanBuilder.ts
@@ -1,0 +1,17 @@
+/**
+ *  Concatenation Builder
+ */
+
+import {GBuilder} from "../core/builder";
+import {GPlugin} from "../core/plugin";
+
+export class GCleanBuilder extends GBuilder {
+  constructor() { super(); }
+
+  build() {
+    let promise = GPlugin.clean(this);
+    if (this.conf.flushStream) return promise;
+  }
+}
+
+export default GCleanBuilder;